Remove output `transition_inner`s
